Start of permissions

+++ bot/cogs/system.py
@@ -2,6 +2,7 @@ import asyncio
 import datetime
 
 from discord.ext import commands
+from discord.utils import get
 
 from bot.reference import *
 import bot.reference as ref
@@ -13,12 +14,24 @@ class Util(commands.Cog):
 
     @commands.command(pass_context=True, name="clear")
     async def clear(self, ctx, number=5):
-        number = int(number)  # Converting the amount of messages to delete to an integer
-        await ctx.channel.purge(limit=number)
-        temp = await ctx.send(f"**@everyone :white_check_mark: {number} message(s) Cleared!**")
-        print(f":wastebasket: Cleared {number} messages from channel: {ctx.channel}")
-        await asyncio.sleep(2.5)
-        await temp.delete()
+        can_send = False
+        roles = {"Sauce Creator": 10, "Sauce Provider": 15, "Admin": 10}
+        for role in ctx.author.roles:
+            if role.name in roles:
+                can_send = True
+
+        if ctx.author.id == "229773126936821760":
+            can_send = True
+
+        if can_send:
+            number = int(number)  # Converting the amount of messages to delete to an integer
+            await ctx.channel.purge(limit=number)
+            temp = await ctx.send(f"**{ctx.author.mention} :white_check_mark: {number} message(s) Cleared!**")
+            print(f":wastebasket: Cleared {number} messages from channel: {ctx.channel}")
+            await asyncio.sleep(2.5)
+            await temp.delete()
+        else:
+            await ctx.channel.send(f"**{ctx.author.mention} You do not have permission to use this!**")
 
     @commands.command(pass_context=True, name="prefix")
     async def prefix(self, ctx, prefix=None):
